+++ demos/src/bin/platformer_2d.rs
@@ -117,6 +117,10 @@ fn main() {
                 "CompoundColliders",
                 tnua_demos_crate::levels_setup::compound_colliders_2d::setup_level,
             )
+            .with(
+                "DynamicBodies",
+                tnua_demos_crate::levels_setup::dynamic_bodies_2d::setup_level,
+            )
     });
     app.add_systems(Startup, setup_player);
     app.add_systems(