+++ demos/src/levels_setup/mod.rs
@@ -1,5 +1,7 @@
 pub mod compound_colliders_2d;
 pub mod compound_colliders_3d;
+pub mod dynamic_bodies_2d;
+pub mod dynamic_bodies_3d;
 pub mod for_2d_platformer;
 pub mod for_3d_platformer;
 mod helper;