+++ rapier3d/CHANGELOG.md
@@ -8,6 +8,7 @@ NOTE: This changelog is shared between bevy-tnua-rapier2d and bevy-tnua-rapier3d
 
 ## [Unreleased]
 ### Added
+- Support for `TnuaNotPlatform`.
 - Support for `TnuaObstacleRadar`.
 - `TnuaSpatialExtRapier2d`/`TnuaSpatialExtRapier3d` - implementation for the
   `TnuaSpatialExt` trait.